+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### Bug fixes
 - Fix `besu -X` unstable options help [#8662](https://github.com/hyperledger/besu/pull/8662)
+- Prevent parsing of invalid ENR records from crashing DNSDaemon [#8368](https://github.com/hyperledger/besu/issues/8368)
 
 ## 25.5.0
 ### Breaking Changes